pyproject.toml
[project]
name = "argumentative-path-rag"
version = "0.1.0"
description = "Query-type-adaptive RAG over argument graphs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # LLM
    "openai>=1.30.0",
    # Data validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # Graph DB
    "neo4j>=5.20.0",
    "networkx>=3.3",
    # Embeddings / retrieval
    "sentence-transformers>=3.0.0",
    "torch>=2.2.0",
    # Document processing
    "tiktoken>=0.7.0",
    "pypdf>=4.2.0",
    # Evaluation
    "ragas>=0.1.14",
    "datasets>=2.19.0",
    # Utilities
    "tenacity>=8.3.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ap_rag"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
